[CBRD-25704] Fix an issue where a core dump occurs during unloaddb when the target_unique_name of a synonym includes the server. (#5651)
http://jira.cubrid.org/browse/CBRD-25704

In issue CBRD-25629, an attempt was made to change the [user_schema] of a synonym to lowercase and to remove the [user_schema] from the target class during an unload operation. However, this resulted in a core dump failure.

After further checking, it was confirmed that the [user_schema] of the target should not be removed for the following reasons. Accordingly, only the [user_schema] of the synonym was modified to lowercase.

Reasons why the [user_schema] of the target should not be removed
1. The owner of the target table is not a local user but a remote user.
2. If the [user_schema] of the target is removed, the user connected from the remote server may be treated as the owner, potentially causing unintended queries to be executed and resulting in failures.